Fix build.sh

+++ Voinova.Alyona/lab1/build.sh
@@ -2,7 +2,7 @@
 
 pidFilePath="/var/run/lab1.pid"
 
-[[ -f pidFilePath ]] || sudo touch "$pidFilePath"
+[[ -f $pidFilePath ]] || sudo touch "$pidFilePath"
 
 sudo chmod 666 "$pidFilePath"
 
@@ -10,7 +10,5 @@ cmake .
 make
 rm CMakeCache.txt Makefile cmake_install.cmake
 FILE=daemon_lab.cbp
-if [[ -f "$FILE" ]]; then
-    rm "$FILE"
-fi
+[[ -f "$FILE" ]] && rm "$FILE"
 rm -r CMakeFiles/
